requirements.txt
-i https://pypi.org/simple
attrs==23.2.0; python_version >= '3.7'
blessed==1.20.0; python_version >= '2.7'
boto3==1.18.62
botocore==1.21.65; python_version >= '3.6'
build==1.0.3
chalice==1.30.0
click==8.1.7; python_version >= '3.7'
inquirer==2.10.1; python_version >= '3.7'
jmespath==0.10.0; python_version >= '2.6' and python_version not in '3.0, 3.1, 3.2, 3.3'
jsonschema==4.21.1
jsonschema-specifications==2023.12.1; python_version >= '3.8'
packaging==23.2; python_version >= '3.7'
pip==23.3.2; python_version >= '3.7'
pyproject-hooks==1.0.0; python_version >= '3.7'
python-dateutil==2.8.2; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
python-editor==1.0.4
pyyaml==6.0.1; python_version >= '3.6'
readchar==4.0.5; python_version >= '3.7'
referencing==0.32.1; python_version >= '3.8'
rpds-py==0.17.1; python_version >= '3.8'
s3transfer==0.5.2; python_version >= '3.6'
setuptools==69.0.3; python_version >= '3.8'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
typing-extensions==4.9.0; python_version >= '3.8'
urllib3==1.26.18; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4, 3.5'
wcwidth==0.2.13
wheel==0.42.0; python_version >= '3.7'
pyjwt==2.3.0; python_version >= '3.6'

